binding/c: autogen mpi_t api
